+++ src/main/java/commands/DeleteCommand.java
@@ -1,10 +1,10 @@
-package Commands;
+package commands;
 
-import Duke.DukeException;
-import Duke.Storage;
-import Duke.TaskList;
-import Duke.Ui;
-import Tasks.Task;
+import data.exception.DukeException;
+import storage.Storage;
+import data.TaskList;
+import ui.Ui;
+import tasks.Task;
 
 public class DeleteCommand extends Command {
