+++ src/main/java/storage/Storage.java
@@ -1,6 +1,7 @@
-package Duke;
+package storage;
 
-import Tasks.*;
+import data.TaskList;
+import tasks.*;
 import java.io.*;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
